update to php 8.4, nette 3, using, doctrine.....

+++ app/Database/Migrations/Version20241027230245.php
@@ -23,7 +23,7 @@ final class Version20241027230245 extends AbstractMigration
     {
         $table = $schema->createTable('public.role');
 
-        $table->addColumn('id', Types::INTEGER)
+        $table->addColumn('id', Types::BIGINT)
             ->setAutoincrement(true)
             ->setComment('ID');
